Further refinement of the sidebar
